+++ packages/core/src/components/Tabs/TabList/types.ts
@@ -1,6 +1,6 @@
 import { HTMLProps, WithThemeProp } from '@medly-components/utils';
 import { TabSize, Variant } from '../types';
-export interface Props extends HTMLProps<HTMLDivElement> {
+export interface TabListProps extends HTMLProps<HTMLDivElement> {
     active: any;
     onChange: (id: any) => void;
 }