+++ packages/core/src/components/ToastContainer/ToastContainer.component.tsx
@@ -4,18 +4,20 @@ import React, { FC } from 'react';
 import Toast from '../Toast';
 import * as Styled from './ToastContainer.styled';
 import { toastStore } from './ToastStore';
-import { Props } from './types';
+import { ToastContainerProps } from './types';
 
-export const ToastContainer: FC<Props> & WithStyle = React.memo(
+export const ToastContainer: FC<ToastContainerProps> & WithStyle = React.memo(
     React.forwardRef((props, ref) => {
         const toasts = useStore(toastStore);
 
         return (
-            <Styled.Container ref={ref} {...props}>
-                {toasts.map(toast => (
-                    <Toast key={toast.id} {...toast} />
-                ))}
-            </Styled.Container>
+            toasts.length > 0 && (
+                <Styled.Container ref={ref} {...props}>
+                    {toasts.map(toast => (
+                        <Toast key={toast.id} {...toast} />
+                    ))}
+                </Styled.Container>
+            )
         );
     })
 );