+++ packages/core/src/components/ToastContainer/types.ts
@@ -1,5 +1,5 @@
 import { HTMLProps } from '@medly-components/utils';
 
-export type Props = HTMLProps<HTMLDivElement> & {
+export type ToastContainerProps = HTMLProps<HTMLDivElement> & {
     position?: 'top-start' | 'top' | 'top-end' | 'right' | 'left' | 'bottom-end' | 'bottom' | 'bottom-start';
 };