+++ packages/loaders/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.11](https://github.com/medly/medly-components/compare/@medly-components/loaders@1.2.10...@medly-components/loaders@1.2.11) (2021-08-11)
+
+**Note:** Version bump only for package @medly-components/loaders
+
+
+
+
+
+## [1.2.10](https://github.com/medly/medly-components/compare/@medly-components/loaders@1.2.9...@medly-components/loaders@1.2.10) (2021-08-11)
+
+**Note:** Version bump only for package @medly-components/loaders
+
+
+
+
+
+## [1.2.9](https://github.com/medly/medly-components/compare/@medly-components/loaders@1.2.8...@medly-components/loaders@1.2.9) (2021-08-05)
+
+**Note:** Version bump only for package @medly-components/loaders
+
+
+
+
+
+## [1.2.8](https://github.com/medly/medly-components/compare/@medly-components/loaders@1.2.7...@medly-components/loaders@1.2.8) (2021-08-05)
+
+**Note:** Version bump only for package @medly-components/loaders
+
+
+
+
+
 ## [1.2.7](https://github.com/medly/medly-components/compare/@medly-components/loaders@1.2.6...@medly-components/loaders@1.2.7) (2021-08-02)
 
 **Note:** Version bump only for package @medly-components/loaders